# Inline lambdas in `across()` expansions