--- a/R/utils.r
+++ b/R/utils.r
@@ -133,3 +133,22 @@ maybe_restart <- function(restart) {
     invokeRestart(restart)
   }
 }
+
+expr_substitute <- function(expr, old, new) {
+  expr <- duplicate(expr)
+  switch(typeof(expr),
+    language = node_walk_replace(node_cdr(expr), old, new),
+    symbol = if (identical(expr, old)) return(new)
+  )
+  expr
+}
+node_walk_replace <- function(node, old, new) {
+  while (!is_null(node)) {
+    switch(
+      typeof(node_car(node)),
+      language = node_walk_replace(node_cdar(node), old, new),
+      symbol = if (identical(node_car(node), old)) node_poke_car(node, new)
+    )
+    node <- node_cdr(node)
+  }
+}