Slightly more performant:

```r
gdf <- tibble(g = seq_len(1e5), x = g) %>% group_by(g)

bench::mark(
  expr = gdf %>% mutate(x = identity(x)),
  across = gdf %>% mutate(across(x, ~ identity(.))),
  iterations = 20
)[1:8]

#> # Before
#>   expression     min  median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
#>   <bch:expr> <bch:t> <bch:t>     <dbl> <bch:byt>    <dbl> <int> <dbl>
#> 1 expr         126ms   131ms      7.22    8.78MB     11.6    20    32
#> 2 across       165ms   172ms      5.46    8.78MB     17.5    20    64

#> # After
#>   expression     min  median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
#>   <bch:expr> <bch:t> <bch:t>     <dbl> <bch:byt>    <dbl> <int> <dbl>
#> 1 expr         126ms   128ms      7.83    8.78MB     86.1     2    22
#> 2 across       129ms   129ms      7.75    9.77MB    171.      1    22
```

This will allow more optimisations in the future when we reintroduce hybrid eval.

The inlining is not 100% correct because we work on the evaluated `.fns` argument. Ideally we would take unevaluated calls and only transform calls to `~` and to `function` (the latter will be important for R 4.1 lambdas). Because of this, we currently have incorrect behaviour with:

```r
f <- local({
  prefix <- "foo"
  ~ paste(prefix, .x)
})

data.frame(x = "x") %>%
  mutate(across(x, f))
#> Error: Problem with `mutate()` column `..1`.
#> ℹ `..1 = across(x, f)`.
#> ✖ object 'prefix' not found
```

Since this is a rather obscure edge case, I think we can go ahead with this PR.